test(notificationController): add getSentNotifications tests
Added four additional tests for getSentNotifications inside controller. I found a bug while working on this test. Because of how logical operators work in JavaScript, the logic that was set up inside the method would always return false. This is not intended behavior so I corrected it by using '&&' instead of '||'. This is on line 81 of the 'notificationsController'